fix unmounted status change issues (hopefully)
Based on https://www.debuggr.io/react-update-unmounted-component/

Referring to this error:

```
index.js:2177 Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
    in PlotLoader (created by MDXCreateElement)
```

Signed-off-by: Gergely Imreh <gergely.imreh@faculty.ai>